+++ src/Database/test.js
@@ -103,7 +103,7 @@ describe('Batch writes', () => {
           m2.body = 'baz1'
         }),
         m3.prepareMarkAsDeleted(),
-        m4.prepareDestroyPermanently()
+        m4.prepareDestroyPermanently(),
       ),
     )
 
@@ -114,27 +114,26 @@ describe('Batch writes', () => {
 
     expect(adapterBatchSpy).toHaveBeenCalledTimes(5)
     expect(adapterBatchSpy).toHaveBeenLastCalledWith([
-      ['create', m6],
-      ['update', m1],
-      ['create', m5],
-      ['update', m2],
-      ['markAsDeleted', m3],
-      ['destroyPermanently', m4],
+      ['create', 'mock_comments', m6._raw],
+      ['update', 'mock_tasks', m1._raw],
+      ['create', 'mock_tasks', m5._raw],
+      ['update', 'mock_comments', m2._raw],
+      ['markAsDeleted', 'mock_tasks', m3.id],
+      ['destroyPermanently', 'mock_comments', m4.id],
     ])
 
     expect(tasksCollectionObserver).toHaveBeenCalledTimes(1)
     expect(commentsCollectionObserver).toHaveBeenCalledTimes(1)
     expect(tasksCollectionObserver).toHaveBeenCalledWith([
-        { record: m1, type: CollectionChangeTypes.updated },
-        { record: m5, type: CollectionChangeTypes.created },
-        { record: m3, type: CollectionChangeTypes.destroyed },
+      { record: m1, type: CollectionChangeTypes.updated },
+      { record: m5, type: CollectionChangeTypes.created },
+      { record: m3, type: CollectionChangeTypes.destroyed },
     ])
     expect(commentsCollectionObserver).toHaveBeenCalledWith([
-        { record: m6, type: CollectionChangeTypes.created },
-        { record: m2, type: CollectionChangeTypes.updated },
-        { record: m4, type: CollectionChangeTypes.destroyed },
+      { record: m6, type: CollectionChangeTypes.created },
+      { record: m2, type: CollectionChangeTypes.updated },
+      { record: m4, type: CollectionChangeTypes.destroyed },
     ])
-
 
     const createdRecords = [m5, m6]
     createdRecords.forEach(record => {
@@ -167,7 +166,7 @@ describe('Batch writes', () => {
     await database.action(() => database.batch(null, model, false, undefined))
 
     expect(adapterBatchSpy).toHaveBeenCalledTimes(1)
-    expect(adapterBatchSpy).toHaveBeenLastCalledWith([['create', model]])
+    expect(adapterBatchSpy).toHaveBeenLastCalledWith([['create', 'mock_tasks', model._raw]])
   })
   it('throws error if attempting to batch records without a pending operation', async () => {
     const { database, tasks } = mockDatabase({ actionsEnabled: true })